requirements.txt
django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pyyaml==6.0
requests==2.28.2
django-cors-headers==3.13.0
pillow==9.4.0
psycopg2==2.9.6
python-dotenv==0.20.0
whitenoise==6.4.0